+++ src/ui/imageproviders.cpp
@@ -17,8 +17,8 @@ QImage getMetatileImage(
         uint16_t metatileId,
         Tileset *primaryTileset,
         Tileset *secondaryTileset,
-        QList<int> layerOrder,
-        QList<float> layerOpacity,
+        const QList<int> &layerOrder,
+        const QList<float> &layerOpacity,
         bool useTruePalettes)
 {
     Metatile* metatile = Tileset::getMetatile(metatileId, primaryTileset, secondaryTileset);
@@ -34,8 +34,8 @@ QImage getMetatileImage(
         Metatile *metatile,
         Tileset *primaryTileset,
         Tileset *secondaryTileset,
-        QList<int> layerOrder,
-        QList<float> layerOpacity,
+        const QList<int> &layerOrder,
+        const QList<float> &layerOpacity,
         bool useTruePalettes)
 {
     QImage metatile_image(16, 16, QImage::Format_RGBA8888);
@@ -43,9 +43,15 @@ QImage getMetatileImage(
         metatile_image.fill(Qt::magenta);
         return metatile_image;
     }
-    metatile_image.fill(Qt::black);
 
     QList<QList<QRgb>> palettes = Tileset::getBlockPalettes(primaryTileset, secondaryTileset, useTruePalettes);
+
+    // We need to fill the metatile image with something so that if any transparent
+    // tile pixels line up across layers we will still have something to render.
+    // The GBA renders transparent pixels using palette 0 color 0. We have this color,
+    // but all 3 games actually overwrite it with black when loading the tileset palettes,
+    // so we have a setting to choose between these two behaviors.
+    metatile_image.fill(projectConfig.setTransparentPixelsBlack ? QColor("black") : QColor(palettes.value(0).value(0)));
 
     QPainter metatile_painter(&metatile_image);
     const int numLayers = 3; // When rendering, metatiles always have 3 layers
@@ -54,7 +60,6 @@ QImage getMetatileImage(
     for (int y = 0; y < 2; y++)
     for (int x = 0; x < 2; x++) {
         int l = layerOrder.size() >= numLayers ? layerOrder[layer] : layer;
-        int bottomLayer = layerOrder.size() >= numLayers ? layerOrder[0] : 0;
 
         // Get the tile to render next
         Tile tile;
@@ -63,25 +68,25 @@ QImage getMetatileImage(
             tile = metatile->tiles.value(tileOffset + (l * 4));
         } else {
             // "Vanilla" metatiles only have 8 tiles, but render 12.
-            // The remaining 4 tiles are rendered either as tile 0 or 0x3014 (tile 20, palette 3) depending on layer type.
+            // The remaining 4 tiles are rendered using user-specified tiles depending on layer type.
             switch (layerType)
             {
             default:
             case METATILE_LAYER_MIDDLE_TOP:
                 if (l == 0)
-                    tile = Tile(0x3014);
+                    tile = Tile(projectConfig.unusedTileNormal);
                 else // Tiles are on layers 1 and 2
                     tile = metatile->tiles.value(tileOffset + ((l - 1) * 4));
                 break;
             case METATILE_LAYER_BOTTOM_MIDDLE:
                 if (l == 2)
-                    tile = Tile();
+                    tile = Tile(projectConfig.unusedTileCovered);
                 else // Tiles are on layers 0 and 1
                     tile = metatile->tiles.value(tileOffset + (l * 4));
                 break;
             case METATILE_LAYER_BOTTOM_TOP:
                 if (l == 1)
-                    tile = Tile();
+                    tile = Tile(projectConfig.unusedTileSplit);
                 else // Tiles are on layers 0 and 2
                     tile = metatile->tiles.value(tileOffset + ((l == 0 ? 0 : 1) * 4));
                 break;
@@ -91,18 +96,14 @@ QImage getMetatileImage(
         QImage tile_image = getTileImage(tile.tileId, primaryTileset, secondaryTileset);
         if (tile_image.isNull()) {
             // Some metatiles specify tiles that are outside the valid range.
-            // These are treated as completely transparent, so they can be skipped without
-            // being drawn unless they're on the bottom layer, in which case we need
-            // a placeholder because garbage will be drawn otherwise.
-            if (l == bottomLayer) {
-                metatile_painter.fillRect(x * 8, y * 8, 8, 8, palettes.value(0).value(0));
-            }
+            // The way the GBA will render these depends on what's in memory (which Porymap can't know)
+            // so we treat them as if they were transparent.
             continue;
         }
 
         // Colorize the metatile tiles with its palette.
         if (tile.palette < palettes.length()) {
-            QList<QRgb> palette = palettes.value(tile.palette);
+            const QList<QRgb> palette = palettes.value(tile.palette);
             for (int j = 0; j < palette.length(); j++) {
                 tile_image.setColor(j, palette.value(j));
             }
@@ -121,12 +122,10 @@ QImage getMetatileImage(
             }
         }
 
-        // The top layer of the metatile has its first color displayed at transparent.
-        if (l != bottomLayer) {
-            QColor color(tile_image.color(0));
-            color.setAlpha(0);
-            tile_image.setColor(0, color.rgba());
-        }
+        // Color 0 is displayed as transparent.
+        QColor color(tile_image.color(0));
+        color.setAlpha(0);
+        tile_image.setColor(0, color.rgba());
 
         metatile_painter.drawImage(origin, tile_image.mirrored(tile.xflip, tile.yflip));
     }
@@ -144,7 +143,7 @@ QImage getTileImage(uint16_t tileId, Tileset *primaryTileset, Tileset *secondary
     return tileset->tiles.value(index, QImage());
 }
 
-QImage getColoredTileImage(uint16_t tileId, Tileset *primaryTileset, Tileset *secondaryTileset, QList<QRgb> palette) {
+QImage getColoredTileImage(uint16_t tileId, Tileset *primaryTileset, Tileset *secondaryTileset, const QList<QRgb> &palette) {
     QImage tileImage = getTileImage(tileId, primaryTileset, secondaryTileset);
     if (tileImage.isNull()) {
         tileImage = QImage(8, 8, QImage::Format_RGBA8888);